requirements.txt
# Web Framework và API
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# MQTT & IoT
paho-mqtt==2.1.0
adafruit-io==2.8.0

# Auth & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
ecdsa==0.19.0
pyasn1==0.6.1
rsa==4.9

# Data Processing
numpy==2.0.2
pandas==2.2.3
scipy==1.13.1

# Visualization
matplotlib==3.9.4
seaborn==0.13.2
contourpy==1.3.0
kiwisolver==1.4.7
pillow==11.1.0

# Config & Environment
python-dotenv==1.0.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2

# Scheduling
APScheduler==3.11.0
schedule==1.2.2
tzlocal==5.3.1

# Utilities
python-dateutil==2.9.0.post0
tabulate==0.9.0
requests==2.32.3
certifi==2025.1.31

# Formatting & Compatibility
charset-normalizer==3.4.1
typing_extensions==4.12.2
annotated-types==0.7.0
anyio==3.7.1
idna==3.10
six==1.17.0
sniffio==1.3.1
h11==0.14.0
click==8.1.8
exceptiongroup==1.2.2

# Date and Time
pytz==2025.2
tzdata==2025.2

# Other
fonttools==4.56.0
pyparsing==3.2.3
packaging==24.2
zipp==3.21.0
importlib_resources==6.5.2
